Seutodat {scFrac}	R Documentation

this is a pre-process function to intersect the shared genes, normalized the input mixture matrix,Y and the single cell marker gene matrix,Y

Description

this is a pre-process function to intersect the shared genes, normalized the input mixture matrix,Y and the single cell marker gene matrix,Y

Usage

x=Seutodat(Seuratobject)

Arguments

X
seurat object

[Package scFrac version 0.1.0 Index]